refactor(travel, place): Domain 형식으로 패키지 구조 개편
