+++ src/main/java/com/yeohangttukttak/api/domain/place/dto/LocationDTO.java
@@ -1,6 +1,6 @@
 package com.yeohangttukttak.api.domain.place.dto;
 
-import com.yeohangttukttak.api.domain.place.Location;
+import com.yeohangttukttak.api.domain.place.entity.Location;
 import com.yeohangttukttak.api.validator.ValidLatitude;
 import com.yeohangttukttak.api.validator.ValidLongitude;
 import lombok.Data;
