+++ src/main/java/com/yeohangttukttak/api/domain/place/entity/PlaceType.java
@@ -1,4 +1,4 @@
-package com.yeohangttukttak.api.domain.place;
+package com.yeohangttukttak.api.domain.place.entity;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.yeohangttukttak.api.domain.ValueBasedEnum;
