+++ src/main/java/com/yeohangttukttak/api/domain/travel/dto/TravelDTO.java
@@ -1,9 +1,12 @@
-package com.yeohangttukttak.api.domain.travel;
+package com.yeohangttukttak.api.domain.travel.dto;
 
 import com.yeohangttukttak.api.domain.file.ImageDTO;
 import com.yeohangttukttak.api.domain.member.MemberDTO;
-import com.yeohangttukttak.api.domain.file.File;
-import com.yeohangttukttak.api.domain.place.Place;
+import com.yeohangttukttak.api.domain.place.entity.Place;
+import com.yeohangttukttak.api.domain.travel.entity.AccompanyType;
+import com.yeohangttukttak.api.domain.travel.entity.Motivation;
+import com.yeohangttukttak.api.domain.travel.entity.TransportType;
+import com.yeohangttukttak.api.domain.travel.entity.Travel;
 import lombok.Data;
 
 import java.time.LocalDate;
