+++ src/main/java/com/yeohangttukttak/api/domain/visit/api/VisitController.java
@@ -3,13 +3,13 @@ package com.yeohangttukttak.api.domain.visit.api;
 import com.yeohangttukttak.api.controller.api.ApiResponse;
 import com.yeohangttukttak.api.domain.member.AgeGroup;
 import com.yeohangttukttak.api.domain.place.dto.LocationDTO;
-import com.yeohangttukttak.api.domain.travel.AccompanyType;
-import com.yeohangttukttak.api.domain.travel.Motivation;
-import com.yeohangttukttak.api.domain.travel.TransportType;
+import com.yeohangttukttak.api.domain.travel.entity.AccompanyType;
+import com.yeohangttukttak.api.domain.travel.entity.Motivation;
+import com.yeohangttukttak.api.domain.travel.entity.TransportType;
 import com.yeohangttukttak.api.domain.visit.dto.VisitSearch;
 import com.yeohangttukttak.api.domain.visit.dto.VisitSearchDTO;
 import com.yeohangttukttak.api.domain.visit.service.VisitSearchService;
-import com.yeohangttukttak.api.domain.place.Location;
+import com.yeohangttukttak.api.domain.place.entity.Location;
 import com.yeohangttukttak.api.validator.ValidValueBasedEnum;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
