+++ src/main/java/com/yeohangttukttak/api/domain/visit/dao/VisitRepository.java
@@ -5,10 +5,10 @@ import com.querydsl.core.types.dsl.BooleanTemplate;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.yeohangttukttak.api.domain.member.AgeGroup;
-import com.yeohangttukttak.api.domain.place.Location;
-import com.yeohangttukttak.api.domain.travel.AccompanyType;
-import com.yeohangttukttak.api.domain.travel.Motivation;
-import com.yeohangttukttak.api.domain.travel.TransportType;
+import com.yeohangttukttak.api.domain.place.entity.Location;
+import com.yeohangttukttak.api.domain.travel.entity.AccompanyType;
+import com.yeohangttukttak.api.domain.travel.entity.Motivation;
+import com.yeohangttukttak.api.domain.travel.entity.TransportType;
 import com.yeohangttukttak.api.domain.visit.entity.Visit;
 import com.yeohangttukttak.api.domain.visit.dto.VisitSearch;
 import jakarta.persistence.EntityManager;
