+++ src/main/java/com/yeohangttukttak/api/domain/visit/dto/VisitSearch.java
@@ -1,13 +1,12 @@
 package com.yeohangttukttak.api.domain.visit.dto;
 
 import com.yeohangttukttak.api.domain.member.AgeGroup;
-import com.yeohangttukttak.api.domain.place.Location;
-import com.yeohangttukttak.api.domain.travel.AccompanyType;
-import com.yeohangttukttak.api.domain.travel.Motivation;
-import com.yeohangttukttak.api.domain.travel.TransportType;
+import com.yeohangttukttak.api.domain.place.entity.Location;
+import com.yeohangttukttak.api.domain.travel.entity.AccompanyType;
+import com.yeohangttukttak.api.domain.travel.entity.Motivation;
+import com.yeohangttukttak.api.domain.travel.entity.TransportType;
 import lombok.Builder;
 import lombok.Data;
-import lombok.ToString;
 
 @Data
 public class VisitSearch {
